修复了 Niva 接口调用超 256 次崩溃的 bug(原项目 issues), 并优化了 http-api

+++ crates/niva/src/app/menu/options.rs
@@ -29,7 +29,7 @@ pub enum MenuItemOption {
         label: NativeLabel,
     },
     Item {
-        id: u8,
+        id: u16,
         label: String,
         enabled: Option<bool>,
         selected: Option<bool>,
